adornerbehavior: refactoring and fixes
okey. so the code had serious issues.
after fixing the multiple children variant, there is no need for the single one. so i removed it.
then i found out that the number of events subscriptions was growing exponentially on each layout reload. that was really not handled well. i also fixed the documentation and code style.